* Implemented EEP A5-08 (incl. Eltako specific profile)
 * Switched form int to byte for internal enocean data representation
 * Using OpenHabs HexUtils
 * Added dimensions to channels and specified units for Number States

+++ src/main/java/org/openhab/binding/openocean/internal/eep/D5_00/D5_00_01.java
@@ -23,8 +23,8 @@ import org.openhab.binding.openocean.internal.messages.ERP1Message;
  */
 public class D5_00_01 extends _1BSMessage {
 
-    final int OPEN = 0 | TeachInBit;
-    final int CLOSED = 1 | TeachInBit;
+    final byte OPEN = 0 | TeachInBit;
+    final byte CLOSED = 1 | TeachInBit;
 
     public D5_00_01() {
         super();
